Вносит последние правка по ДЗ-8 часть1
