+++ js/form.js
@@ -1,3 +1,10 @@
+import {
+  MIN_TITLE_LENGTH,
+  MAX_TITLE_LENGTH,
+  priceOption,
+  roomsOption,
+} from './const.js';
+
 const adForm = document.querySelector('.ad-form');
 const adFormFieldsets = adForm.querySelectorAll('fieldset');
 
@@ -11,19 +18,15 @@ const mapFilter = mapFilters.querySelectorAll('.map__filter');
 const mapFeatures = mapFilters.querySelector('.map__features');
 const mapElements = [...mapFilter, mapFeatures];
 
-const MIN_TITLE_LENGTH = 30;
-const MAX_TITLE_LENGTH = 100;
-
 const pristine = new Pristine(adForm, {
   classTo: 'ad-form__element',
   errorClass: 'ad-form__element--invalid',
   errorTextParent: 'ad-form__element',
 });
-// ------- Валидация заголовка ------- //
+
 const validateTitle = (value) => value.length >= 30 && value.length <= 100;
 
-// не работает переменная с текущей длинной введенного текста ${title.value.length}
-const getTitleErrorMessage = `Длина заголовка должна быть от ${MIN_TITLE_LENGTH} до ${MAX_TITLE_LENGTH} символов. У вас ${title.value.length}`;
+const getTitleErrorMessage = () => `Длина заголовка должна быть от ${MIN_TITLE_LENGTH} до ${MAX_TITLE_LENGTH} символов. Cейчас ${title.value.length} символов`;
 
 pristine.addValidator(
   title,
@@ -31,26 +34,13 @@ pristine.addValidator(
   getTitleErrorMessage,
 );
 
-// ------- Валидация цены ------- //
-/* Не могу понять почему после загрузки страницы,
-не меняя тип жилья, проверка цены не происходит */
-const priceOption = {
-  'bungalow': '0',
-  'flat' : '1000',
-  'hotel' : '3000',
-  'house' : '5000',
-  'palace' : '10000',
-};
-
 const getNumberMinPrice = () => Number(priceOption[type.value]);
 
-const onTypeChange = () => {
-  price.placeholder = priceOption[type.value];
-  price.min = getNumberMinPrice(); //Что такое min.price?
-};
+const validatePrice = () => {
+  price.min = getNumberMinPrice();
 
-// Почему без Number не работает?
-const validatePrice = () => price.value >= Number(price.min);
+  return price.value >= Number(price.min);
+};
 
 const getPriceErrorMessage = () => `Мин.цена за "${type.options[type.selectedIndex].text}" - ${price.min} рублей!`;
 
@@ -60,25 +50,9 @@ pristine.addValidator(
   getPriceErrorMessage,
 );
 
-// ------- Валидация количества комнат ------- //
-/* Не доконца понимаю как работает код.
-Так же не происходит валидация если после загрузкистраницы просто вбить цену за ночь, не меняя тип жилья */
-
-const onCapacityChange = () => {
-  pristine.validate(capacity);
-  pristine.validate(rooms);
-};
-
-const onRoomsChange = () => {
-  pristine.validate(capacity);
-  pristine.validate(rooms);
-};
-
-const roomsOption = {
-  1 : ['1'],
-  2 : ['2', '1'],
-  3 : ['3', '2', '1'],
-  100 : ['0'],
+const onTypeChange = () => {
+  price.placeholder = priceOption[type.value];
+  pristine.valadate(price);
 };
 
 const validateCapacity = () => roomsOption[rooms.value].includes(capacity.value);
@@ -91,6 +65,9 @@ pristine.addValidator(
   getCapacityErrorMessage,
 );
 
+const onCapacityChange = () => pristine.validate(capacity);
+const onRoomsChange = () => pristine.validate(capacity);
+
 const onformSubmit = (evt) => {
   evt.preventDefault();
 
@@ -101,10 +78,6 @@ const onformSubmit = (evt) => {
   }
 };
 
-adForm.addEventListener('submit', onformSubmit);
-
-
-// ------- Включение / выключение формы ------- //
 const makeFormInactive = () => {
   adForm.classList.add('ad-form--disabled');
 
@@ -115,6 +88,7 @@ const makeFormInactive = () => {
   price.removeEventListener('change', onTypeChange);
   capacity.removeEventListener('change', onCapacityChange);
   rooms.removeEventListener('change', onRoomsChange);
+  adForm.removeEventListener('submit', onformSubmit);
 
 };
 
@@ -128,6 +102,7 @@ const makeFormActive = () => {
   price.removeEventListener('change', onTypeChange);
   capacity.addEventListener('change', onCapacityChange);
   rooms.addEventListener('change', onRoomsChange);
+  adForm.addEventListener('submit', onformSubmit);
 };
 
 const makeMapInactive = () => {
